+++ apis/apigateway/v1beta1/zz_usageplan_types.go
@@ -73,10 +73,10 @@ type APIStagesParameters struct {
 type QuotaSettingsInitParameters struct {
 
 	// Maximum number of requests that can be made in a given time period.
-	Limit *float64 `json:"limit,omitempty" tf:"limit,omitempty"`
+	Limit *int64 `json:"limit,omitempty" tf:"limit,omitempty"`
 
 	// Number of requests subtracted from the given limit in the initial time period.
-	Offset *float64 `json:"offset,omitempty" tf:"offset,omitempty"`
+	Offset *int64 `json:"offset,omitempty" tf:"offset,omitempty"`
 
 	// Time period in which the limit applies. Valid values are "DAY", "WEEK" or "MONTH".
 	Period *string `json:"period,omitempty" tf:"period,omitempty"`
@@ -85,10 +85,10 @@ type QuotaSettingsInitParameters struct {
 type QuotaSettingsObservation struct {
 
 	// Maximum number of requests that can be made in a given time period.
-	Limit *float64 `json:"limit,omitempty" tf:"limit,omitempty"`
+	Limit *int64 `json:"limit,omitempty" tf:"limit,omitempty"`
 
 	// Number of requests subtracted from the given limit in the initial time period.
-	Offset *float64 `json:"offset,omitempty" tf:"offset,omitempty"`
+	Offset *int64 `json:"offset,omitempty" tf:"offset,omitempty"`
 
 	// Time period in which the limit applies. Valid values are "DAY", "WEEK" or "MONTH".
 	Period *string `json:"period,omitempty" tf:"period,omitempty"`
@@ -98,11 +98,11 @@ type QuotaSettingsParameters struct {
 
 	// Maximum number of requests that can be made in a given time period.
 	// +kubebuilder:validation:Optional
-	Limit *float64 `json:"limit" tf:"limit,omitempty"`
+	Limit *int64 `json:"limit" tf:"limit,omitempty"`
 
 	// Number of requests subtracted from the given limit in the initial time period.
 	// +kubebuilder:validation:Optional
-	Offset *float64 `json:"offset,omitempty" tf:"offset,omitempty"`
+	Offset *int64 `json:"offset,omitempty" tf:"offset,omitempty"`
 
 	// Time period in which the limit applies. Valid values are "DAY", "WEEK" or "MONTH".
 	// +kubebuilder:validation:Optional
@@ -112,7 +112,7 @@ type QuotaSettingsParameters struct {
 type ThrottleInitParameters struct {
 
 	// The API request burst limit, the maximum rate limit over a time ranging from one to a few seconds, depending upon whether the underlying token bucket is at its full capacity.
-	BurstLimit *float64 `json:"burstLimit,omitempty" tf:"burst_limit,omitempty"`
+	BurstLimit *int64 `json:"burstLimit,omitempty" tf:"burst_limit,omitempty"`
 
 	// Method to apply the throttle settings for. Specfiy the path and method, for example /test/GET.
 	Path *string `json:"path,omitempty" tf:"path,omitempty"`
@@ -124,7 +124,7 @@ type ThrottleInitParameters struct {
 type ThrottleObservation struct {
 
 	// The API request burst limit, the maximum rate limit over a time ranging from one to a few seconds, depending upon whether the underlying token bucket is at its full capacity.
-	BurstLimit *float64 `json:"burstLimit,omitempty" tf:"burst_limit,omitempty"`
+	BurstLimit *int64 `json:"burstLimit,omitempty" tf:"burst_limit,omitempty"`
 
 	// Method to apply the throttle settings for. Specfiy the path and method, for example /test/GET.
 	Path *string `json:"path,omitempty" tf:"path,omitempty"`
@@ -137,7 +137,7 @@ type ThrottleParameters struct {
 
 	// The API request burst limit, the maximum rate limit over a time ranging from one to a few seconds, depending upon whether the underlying token bucket is at its full capacity.
 	// +kubebuilder:validation:Optional
-	BurstLimit *float64 `json:"burstLimit,omitempty" tf:"burst_limit,omitempty"`
+	BurstLimit *int64 `json:"burstLimit,omitempty" tf:"burst_limit,omitempty"`
 
 	// Method to apply the throttle settings for. Specfiy the path and method, for example /test/GET.
 	// +kubebuilder:validation:Optional
@@ -244,7 +244,7 @@ type UsagePlanParameters struct {
 type UsagePlanThrottleSettingsInitParameters struct {
 
 	// The API request burst limit, the maximum rate limit over a time ranging from one to a few seconds, depending upon whether the underlying token bucket is at its full capacity.
-	BurstLimit *float64 `json:"burstLimit,omitempty" tf:"burst_limit,omitempty"`
+	BurstLimit *int64 `json:"burstLimit,omitempty" tf:"burst_limit,omitempty"`
 
 	// The API request steady-state rate limit.
 	RateLimit *float64 `json:"rateLimit,omitempty" tf:"rate_limit,omitempty"`
@@ -253,7 +253,7 @@ type UsagePlanThrottleSettingsInitParameters struct {
 type UsagePlanThrottleSettingsObservation struct {
 
 	// The API request burst limit, the maximum rate limit over a time ranging from one to a few seconds, depending upon whether the underlying token bucket is at its full capacity.
-	BurstLimit *float64 `json:"burstLimit,omitempty" tf:"burst_limit,omitempty"`
+	BurstLimit *int64 `json:"burstLimit,omitempty" tf:"burst_limit,omitempty"`
 
 	// The API request steady-state rate limit.
 	RateLimit *float64 `json:"rateLimit,omitempty" tf:"rate_limit,omitempty"`
@@ -263,7 +263,7 @@ type UsagePlanThrottleSettingsParameters struct {
 
 	// The API request burst limit, the maximum rate limit over a time ranging from one to a few seconds, depending upon whether the underlying token bucket is at its full capacity.
 	// +kubebuilder:validation:Optional
-	BurstLimit *float64 `json:"burstLimit,omitempty" tf:"burst_limit,omitempty"`
+	BurstLimit *int64 `json:"burstLimit,omitempty" tf:"burst_limit,omitempty"`
 
 	// The API request steady-state rate limit.
 	// +kubebuilder:validation:Optional
