+++ apis/apprunner/v1beta1/zz_autoscalingconfigurationversion_types.go
@@ -23,13 +23,13 @@ type AutoScalingConfigurationVersionInitParameters struct {
 	AutoScalingConfigurationName *string `json:"autoScalingConfigurationName,omitempty" tf:"auto_scaling_configuration_name,omitempty"`
 
 	// Maximal number of concurrent requests that you want an instance to process. When the number of concurrent requests goes over this limit, App Runner scales up your service.
-	MaxConcurrency *float64 `json:"maxConcurrency,omitempty" tf:"max_concurrency,omitempty"`
+	MaxConcurrency *int64 `json:"maxConcurrency,omitempty" tf:"max_concurrency,omitempty"`
 
 	// Maximal number of instances that App Runner provisions for your service.
-	MaxSize *float64 `json:"maxSize,omitempty" tf:"max_size,omitempty"`
+	MaxSize *int64 `json:"maxSize,omitempty" tf:"max_size,omitempty"`
 
 	// Minimal number of instances that App Runner provisions for your service.
-	MinSize *float64 `json:"minSize,omitempty" tf:"min_size,omitempty"`
+	MinSize *int64 `json:"minSize,omitempty" tf:"min_size,omitempty"`
 
 	// Key-value map of resource tags.
 	Tags map[string]*string `json:"tags,omitempty" tf:"tags,omitempty"`
@@ -44,7 +44,7 @@ type AutoScalingConfigurationVersionObservation struct {
 	AutoScalingConfigurationName *string `json:"autoScalingConfigurationName,omitempty" tf:"auto_scaling_configuration_name,omitempty"`
 
 	// The revision of this auto scaling configuration.
-	AutoScalingConfigurationRevision *float64 `json:"autoScalingConfigurationRevision,omitempty" tf:"auto_scaling_configuration_revision,omitempty"`
+	AutoScalingConfigurationRevision *int64 `json:"autoScalingConfigurationRevision,omitempty" tf:"auto_scaling_configuration_revision,omitempty"`
 
 	ID *string `json:"id,omitempty" tf:"id,omitempty"`
 
@@ -52,13 +52,13 @@ type AutoScalingConfigurationVersionObservation struct {
 	Latest *bool `json:"latest,omitempty" tf:"latest,omitempty"`
 
 	// Maximal number of concurrent requests that you want an instance to process. When the number of concurrent requests goes over this limit, App Runner scales up your service.
-	MaxConcurrency *float64 `json:"maxConcurrency,omitempty" tf:"max_concurrency,omitempty"`
+	MaxConcurrency *int64 `json:"maxConcurrency,omitempty" tf:"max_concurrency,omitempty"`
 
 	// Maximal number of instances that App Runner provisions for your service.
-	MaxSize *float64 `json:"maxSize,omitempty" tf:"max_size,omitempty"`
+	MaxSize *int64 `json:"maxSize,omitempty" tf:"max_size,omitempty"`
 
 	// Minimal number of instances that App Runner provisions for your service.
-	MinSize *float64 `json:"minSize,omitempty" tf:"min_size,omitempty"`
+	MinSize *int64 `json:"minSize,omitempty" tf:"min_size,omitempty"`
 
 	// Current state of the auto scaling configuration. An INACTIVE configuration revision has been deleted and can't be used. It is permanently removed some time after deletion.
 	Status *string `json:"status,omitempty" tf:"status,omitempty"`
@@ -78,15 +78,15 @@ type AutoScalingConfigurationVersionParameters struct {
 
 	// Maximal number of concurrent requests that you want an instance to process. When the number of concurrent requests goes over this limit, App Runner scales up your service.
 	// +kubebuilder:validation:Optional
-	MaxConcurrency *float64 `json:"maxConcurrency,omitempty" tf:"max_concurrency,omitempty"`
+	MaxConcurrency *int64 `json:"maxConcurrency,omitempty" tf:"max_concurrency,omitempty"`
 
 	// Maximal number of instances that App Runner provisions for your service.
 	// +kubebuilder:validation:Optional
-	MaxSize *float64 `json:"maxSize,omitempty" tf:"max_size,omitempty"`
+	MaxSize *int64 `json:"maxSize,omitempty" tf:"max_size,omitempty"`
 
 	// Minimal number of instances that App Runner provisions for your service.
 	// +kubebuilder:validation:Optional
-	MinSize *float64 `json:"minSize,omitempty" tf:"min_size,omitempty"`
+	MinSize *int64 `json:"minSize,omitempty" tf:"min_size,omitempty"`
 
 	// Region is the region you'd like your resource to be created in.
 	// +upjet:crd:field:TFTag=-
