+++ apis/iot/v1beta1/zz_thinggroup_types.go
@@ -126,7 +126,7 @@ type ThingGroupObservation struct {
 	TagsAll map[string]*string `json:"tagsAll,omitempty" tf:"tags_all,omitempty"`
 
 	// The current version of the Thing Group record in the registry.
-	Version *float64 `json:"version,omitempty" tf:"version,omitempty"`
+	Version *int64 `json:"version,omitempty" tf:"version,omitempty"`
 }
 
 type ThingGroupParameters struct {
